refactor(chat): extract createSkillActivator to its own file
