+++ src/chat-ui.tsx
@@ -12,9 +12,10 @@ import { shownBranch, shownCwd } from "./chat-layout";
 import { createMessageHandler } from "./chat-message-handler";
 import { suggestModels } from "./chat-model-autocomplete";
 import { type PickerState, pickerItemCount } from "./chat-picker";
-import { createPickerHandlers, createSkillActivator } from "./chat-picker-handlers";
+import { createPickerHandlers } from "./chat-picker-handlers";
 import { ChatRowView } from "./chat-row-view";
 import { createMessage, toRows } from "./chat-session";
+import { createSkillActivator } from "./chat-skill-activator";
 import { suggestSlashCommands } from "./chat-slash";
 import { enqueueQueuedMessage, resolveQueueSubmit } from "./chat-submit";
 import { ChatTranscript } from "./chat-transcript";
